+++ native-bridge/src/type_definings.rs
@@ -1,0 +1,6 @@
+#[repr(C)]
+pub enum EllieTypeDefinings {
+    Int,
+    String,
+    Float,
+}