+++ native/src/ellie_type_define.rs
@@ -1,6 +1,0 @@
-
-#[repr(C)]
-pub enum EllieTypeDefines {
-    Int,
-    String,
-}