+++ src/cli_constants.rs
@@ -1,4 +1,4 @@
-pub static ELLIE_VERSION: &'static str = &"3.6.2";
+pub static ELLIE_VERSION: &'static str = &"3.7.3";
 pub static ELLIE_VERSION_NAME: &'static str = &"BeautifulTropicalFish";
 pub static ELLIE_PARSER_VERSION: &'static str = &"1.7.3";
 pub static ELLIE_RUNTIME_VERSION: &'static str = &"0.1.0";